+++ issue.go
@@ -7,8 +7,9 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"github.com/google/go-querystring/query"
 	"net/http"
+
+	"github.com/google/go-querystring/query"
 )
 
 type IssueService struct {
@@ -279,243 +280,6 @@ func (u *IssueService) Assignee(opts *IssueAssigneePutOption) (v *IssueAssigneeP
 	u.client.requestExtHeader(req)
 	req.Header.Set("Content-Type", "application/json")
 	v = new(IssueAssigneePutObject)
-	resp, err = u.client.Do(req, nil)
-	if err != nil {
-		return nil, resp, err
-	}
-	return
-}
-
-type IssueCommentGetOption struct {
-	IssueIdOrKey string `url:"issueIdOrKey"`
-}
-
-type CommentBody struct {
-	Self   string `json:"self"`
-	ID     string `json:"id"`
-	Author struct {
-		Self         string `json:"self"`
-		Name         string `json:"name"`
-		Key          string `json:"key"`
-		EmailAddress string `json:"emailAddress"`
-		AvatarUrls   struct {
-			Four8X48  string `json:"48x48"`
-			Two4X24   string `json:"24x24"`
-			One6X16   string `json:"16x16"`
-			Three2X32 string `json:"32x32"`
-		} `json:"avatarUrls"`
-		DisplayName string `json:"displayName"`
-		Active      bool   `json:"active"`
-		TimeZone    string `json:"timeZone"`
-	} `json:"author"`
-	Body         string `json:"body"`
-	UpdateAuthor struct {
-		Self         string `json:"self"`
-		Name         string `json:"name"`
-		Key          string `json:"key"`
-		EmailAddress string `json:"emailAddress"`
-		AvatarUrls   struct {
-			Four8X48  string `json:"48x48"`
-			Two4X24   string `json:"24x24"`
-			One6X16   string `json:"16x16"`
-			Three2X32 string `json:"32x32"`
-		} `json:"avatarUrls"`
-		DisplayName string `json:"displayName"`
-		Active      bool   `json:"active"`
-		TimeZone    string `json:"timeZone"`
-	} `json:"updateAuthor"`
-	Created    string `json:"created"`
-	Updated    string `json:"updated"`
-	Visibility struct {
-		Type  string `json:"type"`
-		Value string `json:"value"`
-	} `json:"visibility,omitempty"`
-}
-
-type IssueCommentGetObject struct {
-	StartAt    int           `json:"startAt"`
-	MaxResults int           `json:"maxResults"`
-	Total      int           `json:"total"`
-	Comments   []CommentBody `json:"comments"`
-}
-
-func (u *IssueService) CommentGet(opts *IssueCommentGetOption) (v *IssueCommentGetObject, resp *http.Response, err error) {
-	path := fmt.Sprintf("%v/rest/api/2/issue/%v/comment", u.client.endpoint, opts.IssueIdOrKey)
-	// optv, _ := json.Marshal(opts)
-	req, err := http.NewRequest("GET", path, nil)
-	if err != nil {
-		return
-	}
-	u.client.requestExtHeader(req)
-	req.Header.Set("Content-Type", "application/json")
-	v = new(IssueCommentGetObject)
-	resp, err = u.client.Do(req, v)
-	if err != nil {
-		return nil, resp, err
-	}
-	return
-}
-
-type IssueCommentPostOption struct {
-	IssueIdOrKey string          `url:"issueIdOrKey" json:"-"`
-	Body         string          `json:"body"`
-	Visibility   IssueVisibility `json:"visibility,omitempty"`
-}
-
-type IssueVisibility struct {
-	Type  string `json:"type,omitempty"`
-	Value string `json:"value,omitempty"`
-}
-
-type IssueCommentPostObject struct {
-	Self   string `json:"self"`
-	ID     string `json:"id"`
-	Author struct {
-		Self         string `json:"self"`
-		Name         string `json:"name"`
-		Key          string `json:"key"`
-		EmailAddress string `json:"emailAddress"`
-		AvatarUrls   struct {
-			Four8X48  string `json:"48x48"`
-			Two4X24   string `json:"24x24"`
-			One6X16   string `json:"16x16"`
-			Three2X32 string `json:"32x32"`
-		} `json:"avatarUrls"`
-		DisplayName string `json:"displayName"`
-		Active      bool   `json:"active"`
-		TimeZone    string `json:"timeZone"`
-	} `json:"author"`
-	Body         string `json:"body"`
-	UpdateAuthor struct {
-		Self         string `json:"self"`
-		Name         string `json:"name"`
-		Key          string `json:"key"`
-		EmailAddress string `json:"emailAddress"`
-		AvatarUrls   struct {
-			Four8X48  string `json:"48x48"`
-			Two4X24   string `json:"24x24"`
-			One6X16   string `json:"16x16"`
-			Three2X32 string `json:"32x32"`
-		} `json:"avatarUrls"`
-		DisplayName string `json:"displayName"`
-		Active      bool   `json:"active"`
-		TimeZone    string `json:"timeZone"`
-	} `json:"updateAuthor"`
-	Created    string `json:"created"`
-	Updated    string `json:"updated"`
-	Visibility struct {
-		Type  string `json:"type"`
-		Value string `json:"value"`
-	} `json:"visibility"`
-}
-
-func (u *IssueService) CommentPost(opts *IssueCommentPostOption) (v *IssueCommentPostObject, resp *http.Response, err error) {
-	path := fmt.Sprintf("%v/rest/api/2/issue/%v/comment", u.client.endpoint, opts.IssueIdOrKey)
-	optv, _ := json.Marshal(opts)
-	req, err := http.NewRequest("POST", path, bytes.NewBuffer(optv))
-	if err != nil {
-		return
-	}
-	u.client.requestExtHeader(req)
-	req.Header.Set("Content-Type", "application/json")
-	v = new(IssueCommentPostObject)
-	resp, err = u.client.Do(req, nil)
-	if err != nil {
-		return nil, resp, err
-	}
-	return
-}
-
-type IssueTransitionsGetOption struct {
-	IssueIdOrKey string `url:"issueIdOrKey"`
-	TransitionId string `url:"transitionId"`
-}
-
-type IssueTransitionsGetObject struct {
-	Expand      string              `json:"expand"`
-	Transitions []TransitionsObject `json:"transitions"`
-}
-
-type TransitionsObject struct {
-	ID   string `json:"id"`
-	Name string `json:"name"`
-	To   struct {
-		Self           string `json:"self"`
-		Description    string `json:"description"`
-		IconURL        string `json:"iconUrl"`
-		Name           string `json:"name"`
-		ID             string `json:"id"`
-		StatusCategory struct {
-			Self      string `json:"self"`
-			ID        int    `json:"id"`
-			Key       string `json:"key"`
-			ColorName string `json:"colorName"`
-			Name      string `json:"name"`
-		} `json:"statusCategory"`
-	} `json:"to"`
-}
-
-// issue流转
-func (u *IssueService) TransitionsGet(opts *IssueTransitionsGetOption) (v *IssueTransitionsGetObject, resp *http.Response, err error) {
-	path := fmt.Sprintf("%v/rest/api/2/issue/%v/transitions", u.client.endpoint, opts.IssueIdOrKey)
-	if len(opts.TransitionId) > 0 {
-		path = fmt.Sprintf("%v?transitionId=%v", path, opts.TransitionId)
-	}
-	req, err := http.NewRequest("GET", path, nil)
-	if err != nil {
-		return
-	}
-	u.client.requestExtHeader(req)
-	req.Header.Set("Content-Type", "application/json")
-	v = new(IssueTransitionsGetObject)
-	resp, err = u.client.Do(req, v)
-	if err != nil {
-		return nil, resp, err
-	}
-	return
-}
-
-type TTransition struct {
-	ID string `json:"id"`
-}
-
-type TFields struct {
-	Assignee struct {
-		Name string `json:"name"`
-	} `json:"assignee"`
-	Resolution struct {
-		Name string `json:"name"`
-	} `json:"resolution"`
-}
-
-type TUpdate struct {
-	Comment []struct {
-		Add struct {
-			Body string `json:"body"`
-		} `json:"add"`
-	} `json:"comment"`
-}
-
-type IssueTransitionsPostOption struct {
-	IssueIdOrKey string `url:"issueIdOrKey" json:"-"`
-	//Update TUpdate `json:"update"`
-	//Fields TFields`json:"fields"`
-	Transition TTransition `json:"transition"`
-}
-
-type IssueTransitionsPostObject struct {
-}
-
-func (u *IssueService) TransitionsPost(opts *IssueTransitionsPostOption) (v *IssueTransitionsPostObject, resp *http.Response, err error) {
-	path := fmt.Sprintf("%v/rest/api/2/issue/%v/transitions", u.client.endpoint, opts.IssueIdOrKey)
-	optv, _ := json.Marshal(opts)
-	req, err := http.NewRequest("POST", path, bytes.NewBuffer(optv))
-	if err != nil {
-		return
-	}
-	u.client.requestExtHeader(req)
-	req.Header.Set("Content-Type", "application/json")
-	v = new(IssueTransitionsPostObject)
 	resp, err = u.client.Do(req, nil)
 	if err != nil {
 		return nil, resp, err
